+++ domain/model/src/main/java/co/com/crediya/model/user/User.java
@@ -2,7 +2,6 @@ package co.com.crediya.model.user;
 
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -20,5 +19,5 @@ public class User {
     String identityDocument;
     String phoneNumber;
     Integer roleId;
-    BigDecimal baseSalary;
+    Double baseSalary;
 }
