+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/crediya/api/dtos/ResponseRequestDTO.java
@@ -1,16 +1,15 @@
 package co.com.crediya.api.dtos;
 
-import java.math.BigDecimal;
 
-public record ResponseRequestDTO(BigDecimal amount,
+public record ResponseRequestDTO(Double amount,
                                  Integer period,
                                  String email,
                                  Long idState,
                                  Long idLoanType,
                                  String state,
                                  String type,
-                                 BigDecimal basePayment,
+                                 Double basePayment,
                                  String nameClient,
-                                 BigDecimal interestRate,
-                                 BigDecimal totalMonthlyDebtApprovedRequests) {
+                                 Double interestRate,
+                                 Double totalMonthlyDebtApprovedRequests) {
 }
